refactor(study): move single-experiment execution into the study layer
The single-experiment in-process path lived in api/_impl._run_in_process,
which reached into a study-layer private symbol
(from llenergymeasure.study.runner import _save_and_record) - an api layer
importing a study underscore. Move that execution body into a new
study/single.py:run_single_experiment(), making _save_and_record an
intra-package call.

No behavior change: N=1 / n_cycles=1 still runs in-process (not routed
through StudyRunner, which would change to subprocess isolation). _run keeps
the dispatch decision and the single-experiment progress begin/end block and
now delegates to run_single_experiment. The GPU-memory-check test moves to
tests/unit/study/test_single.py; _resolve_gpu_indices test imports point at
its canonical home (device.gpu_info) now that _impl no longer re-exports it.
